+++ Photo Club Hub/PhotoClubHubApp.swift
@@ -55,19 +55,19 @@ extension PhotoClubHubApp {
 
         // warning: following clubs rely on Level 1 file for GPS coordiantes
 
-        // load test member(s) of Fotogroep Bellus Imago TODO
-//        let bellusBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
-//        bellusBackgroundContext.name = "Bellus Imago"
-//        bellusBackgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
-//        bellusBackgroundContext.automaticallyMergesChangesFromParent = true
-//        _ = BellusImagoMembersProvider(bgContext: bellusBackgroundContext)
-//
-//        // load test member(s) of Fotogroep De Gender
-//        let genderBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
-//        genderBackgroundContext.name = "FG de Gender"
-//        genderBackgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
-//        genderBackgroundContext.automaticallyMergesChangesFromParent = true
-//        _ = FotogroepDeGenderMembersProvider(bgContext: genderBackgroundContext)
+        // load test member(s) of Fotogroep Bellus Imago
+        let bellusBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
+        bellusBackgroundContext.name = "Bellus Imago"
+        bellusBackgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
+        bellusBackgroundContext.automaticallyMergesChangesFromParent = true
+        _ = BellusImagoMembersProvider(bgContext: bellusBackgroundContext)
+
+        // load test member(s) of Fotogroep De Gender
+        let genderBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
+        genderBackgroundContext.name = "FG de Gender"
+        genderBackgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
+        genderBackgroundContext.automaticallyMergesChangesFromParent = true
+        _ = FotogroepDeGenderMembersProvider(bgContext: genderBackgroundContext)
 
         // load all current members of Fotogroep Anders
         let andersBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
@@ -76,12 +76,12 @@ extension PhotoClubHubApp {
         andersBackgroundContext.automaticallyMergesChangesFromParent = true
         _ = AndersMembersProvider(bgContext: andersBackgroundContext)
 
-//        // load all current/former members of Fotogroep Waalre
-//        let waalreBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
-//        waalreBackgroundContext.name = "Fotogroep Waalre"
-//        waalreBackgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
-//        waalreBackgroundContext.automaticallyMergesChangesFromParent = true
-//        _ = FotogroepWaalreMembersProvider(bgContext: waalreBackgroundContext)
+        // load all current/former members of Fotogroep Waalre
+        let waalreBackgroundContext = PersistenceController.shared.container.newBackgroundContext()
+        waalreBackgroundContext.name = "Fotogroep Waalre"
+        waalreBackgroundContext.mergePolicy = NSMergePolicy.mergeByPropertyObjectTrump
+        waalreBackgroundContext.automaticallyMergesChangesFromParent = true
+        _ = FotogroepWaalreMembersProvider(bgContext: waalreBackgroundContext)
 
     }